eat even more safely
Check for zip file presence before trying to eat
Also, use shell reboot recovery to use device specific trap commands

+++ envsetup.sh
@@ -630,11 +630,15 @@ function eat()
     if [ "$OUT" ] ; then
         MODVERSION=`sed -n -e'/ro\.modversion/s/^.*CyanogenMod-//p' $OUT/system/build.prop`
         ZIPFILE=$OUT/update-cm-$MODVERSION-signed.zip
+        if [ ! -f $ZIPFILE ] ; then
+            echo "Nothing to eat"
+            return 1
+        fi
         if [ $(adb get-state) != device ] ; then
             echo "No device is online. Waiting for one..."
             adb wait-for-device
         fi
-        echo "Pushing update-cm-$MODVERSION-signed.zip to device"
+        echo "Pushing $ZIPFILE to device"
         if adb push $ZIPFILE /mnt/sdcard/ ; then
             cat << EOF > /tmp/extendedcommand
 ui_print("Nom nom nom nom...");
@@ -642,7 +646,7 @@ install_zip("/sdcard/update-cm-$MODVERSION-signed.zip");
 EOF
             if adb push /tmp/extendedcommand /cache/recovery/ ; then
                 echo "Rebooting into recovery for installation"
-                adb reboot recovery
+                adb shell "reboot recovery && exit"
             fi
             rm /tmp/extendedcommand
         fi
